Better project structure  👨🏻‍💻
I didnt test this yet but it suppose to work
