+++ main.go
@@ -4,32 +4,22 @@ package main
 import (
 	"flag"
 	"fmt"
+	tv "jobgosh/src/timeViewer"
+	types "jobgosh/src/typesAndVars"
+	wTriggers "jobgosh/src/workTriggers"
 	"os"
 	"os/user"
 	"time"
 )
 
-// variables
-var (
-	DOCKER_DEVELOPMENT        = false
-	Version            string = "v0.2"
-	Revision           string = "17/04/2021"
-	logPath            string = "log/"
-	groupPath          string = "group/"
-	homePath           string = "/go/src/work/" // usr.HomeDir
 
-	colorReset string = "\033[0m"
-	colorRed   string = "\033[31m"
-	colorGreen string = "\033[32m"
-	colorCyan  string = "\033[36m"
-)
 
 func main() {
-
+	
 	// disable when u develop
-	if !DOCKER_DEVELOPMENT {
+	if !types.DOCKER_DEVELOPMENT {
 		usr, _ := user.Current()
-		homePath = usr.HomeDir + "/.jobgosh/"
+		types.HomePath = usr.HomeDir + "/.jobgosh/"
 	}
 
 	// init flags
@@ -47,9 +37,9 @@ func main() {
 	} else if *work != "" {
 
 		if *work == "up" || *work == "down" {
-			WorkTriggers(*work) // w command trrigered
+			wTriggers.WorkTriggers(*work) // w command trrigered
 		} else {
-			fmt.Println(string(colorRed), "! ERROR #01, bad hook parameter", string(colorReset))
+			fmt.Println(string(types.ColorRed), "! ERROR #01, bad hook parameter", string(types.ColorReset))
 			os.Exit(0)
 		}
 
@@ -57,10 +47,10 @@ func main() {
 		if *times == "all" {
 			loc, _ := time.LoadLocation("UTC")
 			_now := time.Now().AddDate(0, 0, 1).In(loc).Unix()
-			TimeViewer(0, _now) // t command trrigered
+			tv.TimeViewer(0, _now) // t command trrigered
 		} else {
 
-			fmt.Println(string(colorRed), "! ERROR #02, bad hook parameter", string(colorReset))
+			fmt.Println(string(types.ColorRed), "! ERROR #02, bad hook parameter", string(types.ColorReset))
 			os.Exit(0)
 		}
 	} else if *from != "" && *to != "" {
@@ -69,6 +59,6 @@ func main() {
 		_to, _ := time.Parse("2006/01/02", *to)
 		_UnixedFrom := _from.In(loc).Unix()
 		_UnixedTo := _to.In(loc).Unix()
-		TimeViewer(_UnixedFrom, _UnixedTo) // t command trrigered
+		tv.TimeViewer(_UnixedFrom, _UnixedTo) // t command trrigered
 	}
 }
