+++ src/typesAndVars/const.go
@@ -1,0 +1,8 @@
+package types
+
+const( 
+	ColorReset string = "\033[0m"
+	ColorRed   string = "\033[31m"
+	ColorGreen string = "\033[32m"
+	ColorCyan  string = "\033[36m"
+)
