+++ src/typesAndVars/vars.go
@@ -1,0 +1,12 @@
+package types
+
+var (
+	DOCKER_DEVELOPMENT        = false
+	Version            string = "v0.2"
+	Revision           string = "17/04/2021"
+	LogPath            string = "log/"
+	GroupPath          string = "group/"
+	HomePath           string = "/go/src/work/" // usr.HomeDir
+
+	
+)
